Move processWebhookMessage function to rc-lib package
