+++ packages/rocketchat-lib/server/functions/index.js
@@ -24,3 +24,4 @@ export { unarchiveRoom } from './unarchiveRoom';
 export { updateMessage } from './updateMessage';
 export { validateCustomFields } from './validateCustomFields';
 export { generateUsernameSuggestion } from './getUsernameSuggestion';
+export { processWebhookMessage } from './processWebhookMessage';
